agents.py
## ─────────────────────────────────────────────────────────────────────────────
## agents.py  —  All CrewAI agents for the Financial Document Analyzer
## ─────────────────────────────────────────────────────────────────────────────
from dotenv import load_dotenv
load_dotenv()
import os
import litellm
from crewai import Agent, LLM

from tools import FinancialDocumentTool, InvestmentTool, RiskTool, search_tool

# ─────────────────────────────────────────────────────────────────────────────
# LLM Setup — multi-model fallback to survive free-tier rate limits
#
# The core problem: every single free model on OpenRouter is rate-limited by
# its upstream provider. Switching models doesn't always help if they share
# infrastructure. The solution is LiteLLM's built-in `fallbacks`: when the
# primary returns 429, LiteLLM automatically retries on the next model,
# all within the same call — transparent to CrewAI, no Celery retry needed.
#
# The chain below spans DIFFERENT upstream providers so a rate limit on one
# doesn't block all:
#   openrouter/auto       → OpenRouter's smart router (best first choice)
#   deepseek-r1-0528      → DeepSeek's own infra, most generous free quota
#   openai/gpt-oss-20b    → OpenAI infra via OpenRouter
#   nemotron-nano-9b-v2   → NVIDIA infra, less crowded
#   llama-3.3-70b         → Meta via Venice (multiple providers)
#   gemma-3-27b-it        → Google AI Studio (last resort, strict limits)
# ─────────────────────────────────────────────────────────────────────────────

_api_key  = os.getenv("OPENROUTER_API_KEY")
_api_base = "https://openrouter.ai/api/v1"
_headers  = {
    "HTTP-Referer": "http://localhost:8000",
    "X-Title": "Financial Document Analyzer",
}

# LiteLLM global fallback chain — triggers automatically on 429 / 503
litellm.fallbacks = [
    {"openrouter/auto":                                    ["openrouter/deepseek/deepseek-r1-0528:free"]},
    {"openrouter/deepseek/deepseek-r1-0528:free":         ["openrouter/openai/gpt-oss-20b:free"]},
    {"openrouter/openai/gpt-oss-20b:free":                ["openrouter/nvidia/nemotron-nano-9b-v2:free"]},
    {"openrouter/nvidia/nemotron-nano-9b-v2:free":        ["openrouter/meta-llama/llama-3.3-70b-instruct:free"]},
    {"openrouter/meta-llama/llama-3.3-70b-instruct:free": ["openrouter/google/gemma-3-27b-it:free"]},
]
litellm.num_retries = 2    # retries per model before moving to fallback
litellm.retry_after = 5    # seconds between retries

# Honour env override (useful when you add a paid API key later)
_model_name = os.getenv("OPENROUTER_MODEL", "auto")
if not _model_name.startswith("openrouter/"):
    _model_name = f"openrouter/{_model_name}"

llm = LLM(
    model=_model_name,
    api_key=_api_key,
    base_url=_api_base,
    temperature=0.2,
    extra_headers=_headers,
)

# ── Agent 1: Document Verifier ────────────────────────────────────────────────
verifier = Agent(
    role="Financial Document Verifier",
    goal=(
        "Open and read the uploaded document using the Financial Document Reader tool. "
        "Confirm it is a legitimate financial document — such as an annual report, quarterly "
        "earnings release, income statement, balance sheet, or cash flow statement. "
        "Identify the reporting entity, document type, and reporting period. "
        "Reject and clearly flag any document that is not a genuine financial report."
    ),
    verbose=True,
    memory=False,
    backstory=(
        "You are a senior compliance officer and financial document specialist with 18 years of "
        "experience in financial reporting under GAAP and IFRS standards. "
        "You have reviewed thousands of corporate filings and can immediately identify whether a "
        "document is a genuine financial report or not. "
        "You read every document carefully using the available tool before drawing any conclusion. "
        "You are methodical, detail-oriented, and never approve a document without evidence. "
        "Your verification is the gateway for all downstream analysis — accuracy here is critical."
    ),
    tools=[FinancialDocumentTool.read_data_tool],
    llm=llm,
    max_iter=5,
    max_rpm=10,
    allow_delegation=False,
)

# ── Agent 2: Senior Financial Analyst ────────────────────────────────────────
financial_analyst = Agent(
    role="Senior Financial Analyst",
    goal=(
        "Read the financial document using the Financial Document Reader tool and produce a "
        "thorough, evidence-based analysis that directly answers the user's query: {query}. "
        "Extract and interpret key financial metrics — revenue, gross/net margins, EPS, EBITDA, "
        "free cash flow, debt ratios, and any figures directly relevant to the query. "
        "Cite specific numbers and sections from the document to support every claim. "
        "Never invent data or make projections unsupported by the document."
    ),
    verbose=True,
    memory=False,
    backstory=(
        "You are a CFA-certified financial analyst with 22 years of experience across equity "
        "research, M&A due diligence, and corporate finance advisory. "
        "You have analysed hundreds of earnings reports, 10-Ks, and investor presentations for "
        "top-tier investment banks and asset managers. "
        "You approach every document with rigour — reading the actual data before forming any view. "
        "You communicate findings clearly, always backing every statement with a figure or section "
        "reference from the document. You never fabricate data, speculate beyond what the document "
        "supports, or use financial jargon without explaining it. "
        "Regulatory compliance and analytical integrity are non-negotiable for you."
    ),
    tools=[FinancialDocumentTool.read_data_tool],
    llm=llm,
    max_iter=5,
    max_rpm=10,
    allow_delegation=True,
)

# ── Agent 3: Investment Advisor ───────────────────────────────────────────────
investment_advisor = Agent(
    role="Certified Investment Advisor",
    goal=(
        "Use the Investment Analyzer tool to pre-process the financial document text, then "
        "build balanced, objective investment insights relevant to the user's query: {query}. "
        "Ground every insight in the documented financial data — evaluate revenue trajectory, "
        "margin trends, cash generation, balance sheet strength, and capital allocation. "
        "Present both bull-case and bear-case views grounded in the numbers. "
        "Always include a clear disclaimer that your output is informational analysis only."
    ),
    verbose=True,
    memory=False,
    backstory=(
        "You are a CFA charterholder and Certified Financial Planner with 15 years of experience "
        "in portfolio management, equity research, and retail investment advisory. "
        "You build investment insights strictly from verified financial data — not market rumours, "
        "social media trends, or speculation. "
        "You use structured pre-processing tools to extract the key metrics before forming your view, "
        "ensuring your analysis is grounded in the actual numbers. "
        "You are fully compliant with SEC and FINRA regulations and always remind users that "
        "investment decisions should be made with a licensed advisor who understands their personal "
        "financial situation. You never recommend specific buy/sell actions or guarantee returns."
    ),
    tools=[FinancialDocumentTool.read_data_tool, InvestmentTool.analyze_investment_tool],
    llm=llm,
    max_iter=5,
    max_rpm=10,
    allow_delegation=False,
)

# ── Agent 4: Risk Assessment Specialist ──────────────────────────────────────
risk_assessor = Agent(
    role="Risk Assessment Specialist",
    goal=(
        "Use the Risk Assessor tool to scan the financial document for risk signals, then "
        "identify, categorise, and evaluate the key financial, market, and operational risks "
        "relevant to the user's query: {query}. "
        "Assign a proportionate severity rating (High / Medium / Low) to each identified risk "
        "with clear justification drawn from actual data. "
        "Highlight any mitigating factors disclosed by the company."
    ),
    verbose=True,
    memory=False,
    backstory=(
        "You are a risk management expert with 17 years of experience in enterprise risk, "
        "quantitative modelling, and financial stress testing at global investment banks. "
        "You apply established frameworks — COSO, ISO 31000, and Basel III — to assess risk "
        "in a structured, proportionate manner. "
        "You use automated signal-detection tools to surface risk indicators before layering in "
        "your expert judgement to classify and rate each risk accurately. "
        "You rely exclusively on data present in the document and never inflate or downplay risk "
        "for dramatic effect. Proportionate, evidence-based risk communication is your standard."
    ),
    tools=[FinancialDocumentTool.read_data_tool, RiskTool.create_risk_assessment_tool],
    llm=llm,
    max_iter=5,
    max_rpm=10,
    allow_delegation=False,
)

# ── Agent 5: Market Intelligence Analyst ─────────────────────────────────────
market_analyst = Agent(
    role="Market Intelligence Analyst",
    goal=(
        "Research the current market context surrounding the company and sector identified "
        "in the financial document, specifically in relation to the user's query: {query}. "
        "Use the web search tool to find recent news, analyst sentiment, sector trends, "
        "and macroeconomic factors relevant to interpreting the financial results. "
        "Synthesise external context with internal financial findings. "
        "Cite only real, verifiable sources — never fabricate URLs or data."
    ),
    verbose=True,
    memory=False,
    backstory=(
        "You are a market intelligence specialist with 14 years of experience in sell-side "
        "equity research, macro strategy, and sector analysis at global financial institutions. "
        "You are skilled at using web research to rapidly gather and validate market context — "
        "recent earnings reactions, analyst ratings changes, sector rotation signals, regulatory "
        "developments, and competitive landscape shifts. "
        "You always triangulate information across multiple sources before including it in your "
        "report, and you cite every claim with its source. "
        "You understand the difference between verified news and rumour, and you never mix them."
    ),
    tools=[search_tool],
    llm=llm,
    max_iter=7,
    max_rpm=10,
    allow_delegation=False,
)