requirements.txt
# ══════════════════════════════════════════════════════════════════════════════
# Financial Document Analyzer — Dependencies
# Python 3.11 | Uses uv in Dockerfile for dependency resolution
# ══════════════════════════════════════════════════════════════════════════════

# ── Core AI framework ──────────────────────────────────────────────────────────
crewai==0.130.0
crewai-tools==0.47.1

# ── LangChain sub-packages only (no bare "langchain" meta-package) ─────────────
langchain-community>=0.3.12   # tools.py  → PyPDFLoader
langchain-openai>=0.2.14      # agents.py → ChatOpenAI

# Uncomment ONE if not using OpenAI:
# langchain-google-genai>=2.0.7
# langchain-anthropic>=0.3.3

# ── OpenAI SDK ─────────────────────────────────────────────────────────────────
openai>=1.54.3,<2.0.0
tiktoken>=0.8.0

# ── PDF parsing ────────────────────────────────────────────────────────────────
pypdf>=5.0.0
psycopg2-binary==2.9.9

# ── Web API ────────────────────────────────────────────────────────────────────
fastapi==0.110.3
uvicorn[standard]==0.30.1
python-multipart==0.0.9
httpx==0.27.2

# ── Queue worker ───────────────────────────────────────────────────────────────
celery==5.4.0
redis==5.0.8
flower

# ── Database ───────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.31
alembic==1.13.1
# psycopg2-binary   # uncomment for PostgreSQL in production

# ── Utilities ──────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
pydantic[email]>=2.7.4